# gi_service/contract/rest/organisation_api_contract.yaml
openapi: 3.1.0

info:
  title: OpenGIN Service API - Organization
  description: This document gives the structure of APIs for the OpenGIN Organization Services.
  version: 1.0.0

servers:
  - url: http://localhost:8081

tags:
  - name: Organization
    description: APIs for organization
  - name: Data
    description: APIs for data

paths:
  /v1/all-presidents:
    get:
      tags:
        - Organization
      summary: Get all president list with terms
      description: Returns the list of presidents and terms.
      operationId: getPresidents
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      $ref: "#/components/schemas/President"
              example:
                body:
                  - name: Anura Kumara Dissanayake
                    id: 2403-03-01_cit_1
                    terms:
                      - startDate:
                          date: 2024-09-23T00:00:00Z
                        endDate:
                          date: ""
        "500":
          description: Server Error

  /v1/all-gazettes:
    get:
      tags:
        - Organization
      summary: Get all published gazettes
      description: Returns a list of sorted date objects that are related to gazette published.
      operationId: getGazettes
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      required:
                        - gazetteId
                        - publishedDate
                      properties:
                        gazetteId:
                          type: string
                        publishedDate:
                          type: object
                          required:
                            - date
                          properties:
                            date:
                              type: string
              example:
                body:
                  - gazetteId: 2411-09
                    publishedDate:
                      date: 2024-11-18T00:00:00Z
                  - gazetteId: 2411-10
                    publishedDate:
                      date: 2024-11-18T00:00:00Z
        "500":
          description: Server Error

  /v1/organisation/prime-minister:
    post:
      tags:
        - Organization
      summary: Get prime minister Data
      description: Returns the prime minister details based on the given date.
      operationId: getPrimeMinister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Date"
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: object
                    required:
                      - name
                      - id
                    properties:
                      name:
                        type: string
                      id:
                        type: string
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/organisation/cabinet-flow/{president_id}:
    post:
      tags:
        - Organization
      summary: Get cabinet flow for a president.
      description: Returns cabinet flow data for a president across up to three given dates.
      operationId: getCabinetFlow
      parameters:
        - name: president_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: List of dates (max 3) for which to compute the cabinet flow.
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      type: object
                      required:
                        - name
                        - time
                      properties:
                        name:
                          type: string
                        time:
                          type: string
                  links:
                    type: array
                    items:
                      type: object
                      required:
                        - source
                        - target
                        - value
                      properties:
                        source:
                          type: integer
                        target:
                          type: integer
                        value:
                          type: integer
              example:
                nodes:
                  - name: "min_01"
                    time: "2024-09-23T00:00:00Z"
                  - name: "min_02"
                    time: "2024-11-18T00:00:00Z"
                links:
                  - source: 0
                    target: 1
                    value: 3
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/person-history/{personId}:
    get:
      tags:
        - Organization
      summary: get person history
      description: Returns the history of the person.
      operationId: getPersonHistory
      parameters:
        - name: personId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      properties:
                        startDate:
                          type: object
                          required:
                            - date
                          properties:
                            date:
                              type: string
                        endDate:
                          type: object
                          required:
                            - date
                          properties:
                            date:
                              type: string
                        isPresident:
                          type: boolean
                        portfolio:
                          type: object
                          required:
                            - name
                            - id
                          properties:
                            name:
                              type: string
                            id:
                              type: string
                            ministers:
                              type: array
                              items:
                                type: object
                                required:
                                  - name
                                  - id
                                properties:
                                  name:
                                    type: string
                                  id:
                                    type: string
              example:
                body:
                  - startDate:
                      date: 2024-11-18
                    endDate:
                      date: Present
                    isPresident: true
                    portfolio:
                      name: Minister of Digital Economy
                      id: 2434-min-1
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/department-history/{departmentId}:
    get:
      tags:
        - Organization
      summary: Get department history.
      description: Returns the history of the department.
      operationId: getDepartmentHistory
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      properties:
                        startDate:
                          type: object
                          required:
                            - date
                          properties:
                            date:
                              type: string
                        endDate:
                          type: object
                          required:
                            - date
                          properties:
                            date:
                              type: string
                        portfolio:
                          type: object
                          required:
                            - name
                            - id
                          properties:
                            name:
                              type: string
                            id:
                              type: string
                            ministers:
                              type: array
                              items:
                                type: object
                                required:
                                  - name
                                  - id
                                properties:
                                  name:
                                    type: string
                                  id:
                                    type: string
              example:
                body:
                  - startDate:
                      date: 2024-11-18
                    endDate:
                      date: Present
                    portfolio:
                      name: Minister of Plantation and Community Infrastructure
                      id: 2434-min-2
                      ministers:
                        - name: Samantha Vidyarathna
                          id: 2405-cit-4
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/active-portfolio-list:
    post:
      tags:
        - Organization
      summary: Get active portfolio list.
      description: Returns a list of portfolios under a given president and a given date.
      operationId: getActivePortfolioList
      parameters:
        - name: presidentId
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: This request body corresponds to the date used to retrive the list of active portfolios.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Date"
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: object
                    properties:
                      portfolioList:
                        type: array
                        items:
                          type: object
                          required:
                            - name
                            - id
                          properties:
                            name:
                              type: string
                            id:
                              type: string
                            isNew:
                              type: boolean
                            ministers:
                              type: array
                              items:
                                type: object
                                required:
                                  - name
                                  - id
                                properties:
                                  name:
                                    type: string
                                  id:
                                    type: string
                                  isPresident:
                                    type: boolean
                                  isNew:
                                    type: boolean
                      activeMinistries:
                        type: integer
                      newMinistries:
                        type: integer
                      newMinisters:
                        type: integer
                      ministriesUnderPresident:
                        type: integer
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/gazettes-by-president/{presidentId}:
    post:
      tags:
        - Organization
      summary: Get a list of gazettes for a given time range and president.
      description: Returns a list of gazettes for a given time range and president, sort the values according to the date.
      operationId: getGazettesByPresident
      parameters:
        - name: presidentId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        description: This request body corresponds to the date range used to retrieve the gazette list for the given president term and date range.
        content:
          application/json:
            schema:
              type: object
              description: Required both startDate and endDate for the range.
              properties:
                startDate:
                  type: object
                  required:
                    - date
                  properties:
                    date:
                      type: string
                endDate:
                  type: object
                  required:
                    - date
                  properties:
                    date:
                      type: string
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      additionalProperties:
                        type: array
                        items:
                          type: object
                          required:
                            - gazetteId
                            - publishedDate
                          properties:
                            gazetteId:
                              type: string
                            publishedDate:
                              type: object
                              required:
                                - date
                              properties:
                                date:
                                  type: string
              example:
                body:
                  - "2025-04-23":
                      - gazetteId: 2411-09
                        publishedDate:
                          date: 2024-11-23T00:00:00Z
                      - gazetteId: 2411-10
                        publishedDate:
                          date: 2024-11-23T00:00:00Z
                  - "2025-04-24":
                      - gazetteId: 2411-10
                        publishedDate:
                          date: 2024-11-24T00:00:00Z
                  - "2025-04-25":
                      - gazetteId: 2411-10
                        publishedDate:
                          date: 2024-11-25T00:00:00Z
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/presidents-by-range:
    post:
      tags:
        - Organization
      summary: Get a list of presidents for a given time range.
      description: Returns a list of presidents for a given time range.
      operationId: getPresidentsForRange
      requestBody:
        required: true
        description: This request body corresponds to the date range used to retrieve the president list.
        content:
          application/json:
            schema:
              type: object
              description: required both startDate and endDate for the range
              properties:
                startDate:
                  type: object
                  required:
                    - date
                  properties:
                    date:
                      type: string
                endDate:
                  type: object
                  required:
                    - date
                  properties:
                    date:
                      type: string
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      allOf:
                        - type: object
                          allOf:
                            - type: object
                              required:
                                - name
                                - id
                              properties:
                                name:
                                  type: string
                                id:
                                  type: string
                        - type: object
                          required:
                            - terms
                          properties:
                            terms:
                              type: array
                              items:
                                type: object
                                required:
                                  - startDate
                                  - endDate
                                properties:
                                  startDate:
                                    type: object
                                    required:
                                      - date
                                    properties:
                                      date:
                                        type: string
                                  endDate:
                                    type: object
                                    required:
                                      - date
                                    properties:
                                      date:
                                        type: string
              example:
                body:
                  - name: Anura Kumara Dissanayake
                    id: 2403-03-01_cit_1
                    terms:
                      - startDate:
                          date: 2024-09-23T00:00:00Z
                        endDate:
                          date: ""
        "500":
          description: Server Error

  /v1/departments-by-portfolio/{portfolio_id}:
    post:
      tags:
        - Organization
      summary: Get active departments for a portfolio.
      description: Returns a list of departments under a given portfolio and a given date.
      operationId: getDepartmentsByPortfolio
      parameters:
        - name: portfolio_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: This request body corresponds to the date used to retrive the list of active departments.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Date"
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      required:
                        - name
                        - id
                      properties:
                        name:
                          type: string
                        id:
                          type: string
                        isNew:
                          type: boolean
                        hasData:
                          type: boolean

              example:
                body:
                  - totalDepartments: 2
                    newDepartments: 2
                    departmentList:
                      - name: National Police Academy
                        id: 2187-27_dep_4
                        isNew: false
                        hasData: true
                      - name: Assets vested to the Secretary to the Treasury, under the Act, No. 43 of 2011
                        id: 2153-12_dep_437
                        isNew: false
                        hasData: false
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

  /v1/persons-by-portfolio/{portfolioId}:
    post:
      tags:
        - Organization
      summary: Get active persons for a portfolio.
      description: Returns a list of persons under a given portfolio and a given date.
      operationId: getPersonsByMinistry
      parameters:
        - name: portfolioId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        description: This request body corresponds to the date used to retrieve the list of active persons.
        content:
          application/json:
            schema:
              type: object
              description: required date
              properties:
                activeDate:
                  type: object
                  required:
                    - date
                  properties:
                    date:
                      type: string
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  body:
                    type: array
                    items:
                      type: object
                      required:
                        - name
                        - id
                      properties:
                        name:
                          type: string
                        id:
                          type: string
                        isPresident:
                          type: boolean
                        isNew:
                          type: boolean

              example:
                body:
                  - name: Anura Kumara Dissanayake
                    id: 2403-03_cit_1
                    isPresident: true
                    isNew: false
                  - name: Gotabaya Rajapaksa
                    id: 2149-34_cit_1
                    isPresident: false
                    isNew: false

        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error

components:
  schemas:
    Date:
      type: object
      required:
        - date
      properties:
        date:
          type: string

    Term:
      type: object
      required:
        - startDate
        - endDate
      properties:
        startDate:
          type: object
          required:
            - date
          properties:
            date:
              type: string
        endDate:
          type: object
          required:
            - date
          properties:
            date:
              type: string

    President:
      allOf:
        - type: object
          allOf:
            - type: object
              required:
                - name
                - id
              properties:
                name:
                  type: string
                id:
                  type: string
        - type: object
          required:
            - terms
          properties:
            terms:
              type: array
              items:
                type: object
                required:
                  - startDate
                  - endDate
                properties:
                  startDate:
                    type: object
                    required:
                      - date
                    properties:
                      date:
                        type: string
                  endDate:
                    type: object
                    required:
                      - date
                    properties:
                      date:
                        type: string

    Person:
      type: object
      required:
        - name
        - id
      properties:
        name:
          type: string
        id:
          type: string

    Gazette:
      type: object
      required:
        - gazetteId
        - publishedDate
      properties:
        gazetteId:
          type: string
        publishedDate:
          type: object
          required:
            - date
          properties:
            date:
              type: string

    Portfolio:
      type: object
      required:
        - name
        - id
      properties:
        name:
          type: string
        id:
          type: string
        ministers:
          type: array
          items:
            type: object
            required:
              - name
              - id
            properties:
              name:
                type: string
              id:
                type: string

    Department:
      type: object
      required:
        - name
        - id
      properties:
        name:
          type: string
        id:
          type: string

  /v1/organisation/department-history/{department_id}:
    get:
      tags:
        - Organization
      summary: Get department history timeline.
      description: Returns a timeline of a department including ministry relations and ministers.
      operationId: getDepartmentHistoryTimeline
      parameters:
        - name: department_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successfully returned
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    ministry_id:
                      type: string
                    ministry_name:
                      type: string
                    minister_id:
                      type: string
                    minister_name:
                      type: string
                    period:
                      type: string
              example:
                - ministry_id: "min_01"
                  ministry_name: "Ministry of Finance"
                  minister_id: "pers_01"
                  minister_name: "John Doe"
                  period: "2020 Jan 01 - 2021 Jan 01"
        "400":
          description: Bad request
        "404":
          description: Not found
        "500":
          description: Server Error